+++ mod1_algorithm/mod1_plane/set.cpp
@@ -1,9 +1,9 @@
 #include "mod1_plane.h"
 
 void					mod1_plane::set(
-						const mod1_point2<int> &min,
-						const mod1_point2<int> &max,
-						const int &delta)
+						const mod1_point2<float> &min,
+						const mod1_point2<float> &max,
+						const float &delta)
 {
 	is_set = true;
 	this->size = (max - min) / delta;
